+++ s/shiny/context.ts
@@ -1,13 +1,13 @@
 
 import {css, CSSResultGroup} from "lit"
-import {Flat} from "../flatstate/flat.js"
-import {SignalTower} from "../signals/tower.js"
-import {WatchTower} from "../watch/watch_tower.js"
 
 export class Context {
-	theme: CSSResultGroup = css``
-	flat = new Flat()
-	signals = new SignalTower()
-	watch = new WatchTower()
+	theme: CSSResultGroup = css`
+		* {
+			margin: 0;
+			padding: 0;
+			box-sizing: border-box;
+		}
+	`
 }