+++ s/shiny/state.ts
@@ -1,0 +1,9 @@
+
+import {Flat} from "../flatstate/flat.js"
+import {SignalTower} from "../signals/tower.js"
+import {WatchTower} from "../watch/watch_tower.js"
+
+export const flat = new Flat()
+export const signals = new SignalTower()
+export const watch = new WatchTower()
+